pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = "test project for dimatech"
authors = ["'Kimiyori' <'maksimkalin@mail.ru'>"]
readme = "README.md"
packages = [{include = "test_project"}]

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^22.9.1"
environs = "^9.5.0"
alembic = "^1.8.1"
asyncpg = "^0.27.0"
uvicorn = "^0.19.0"
pydantic = "^1.10.2"
dependency-injector = "^4.40.0"
redis = "^4.3.4"
sanic-jwt = "^1.8.0"
pycryptodome = "^3.15.0"
sqlalchemy = {extras = ["mypy"], version = "^1.4.43"}

[tool.poetry.group.test.dependencies]
faker = "^15.3.1"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.1"

[tool.poetry.group.linters.dependencies]
sanic-testing = "^22.9.0"
black = "^22.10.0"
flake8 = "^5.0.4"
pylint = "^2.15.5"
types-redis = "^4.3.21.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
